+++ springapp/src/main/java/com/examly/springapp/controller/AdmissionController.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import com.examly.springapp.models.*;
 import java.util.*;
 import com.examly.springapp.service.*;
+import com.examly.springapp.service.admin.AdminService;
 
 @RestController
 @CrossOrigin(origins = "*", allowedHeaders = "*")
