+++ springapp/src/main/java/com/examly/springapp/service/admin/AdminServiceImpl.java
@@ -1,17 +1,25 @@
-package com.examly.springapp.service;
+package com.examly.springapp.service.admin;
 
-import com.examly.springapp.repository.*;
-import com.examly.springapp.models.*;
-import org.springframework.stereotype.Service;
-import org.springframework.beans.factory.annotation.*;
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
 
-import java.util.*;
+import com.examly.springapp.models.AdmissionModel;
+import com.examly.springapp.models.CourseModel;
+import com.examly.springapp.models.InstituteModel;
+import com.examly.springapp.models.StudentModel;
+import com.examly.springapp.repository.AdmissionRepo;
+import com.examly.springapp.repository.CourseRepository;
+import com.examly.springapp.repository.InstituteRepository;
+import com.examly.springapp.repository.StudentRepository;
 
 @Service
-public class AdminService {
-
+public class AdminServiceImpl implements AdminService {
+    
     @Autowired
     private StudentRepository studentRepository;
 
@@ -24,10 +32,12 @@ public class AdminService {
     @Autowired
     private AdmissionRepo admissionRepo;
 
+    @Override
     public void addStudent(StudentModel studentModel) {
         studentRepository.save(studentModel);
     }
 
+    @Override
     public String editStudent(int studentId, StudentModel updatedStudent) {
         Optional<StudentModel> stu = studentRepository.findByStudentId(studentId);
         if (stu.isPresent()) {
@@ -59,6 +69,7 @@ public class AdminService {
         return "Student not found";
     }
 
+    @Override
     public String deleteStudent(Integer id) {
         Optional<StudentModel> studentmodel = studentRepository.findById(id);
         if (studentmodel.isPresent()) {
@@ -68,6 +79,7 @@ public class AdminService {
         return "Student not Found";
     }
 
+    @Override
     public StudentModel getStudent(Integer id) {
         Optional<StudentModel> studentmodel = studentRepository.findById(id);
         if (studentmodel.isPresent()) {
@@ -76,14 +88,17 @@ public class AdminService {
         return null;
     }
 
+    @Override
     public List<StudentModel> getStudent() {
         return studentRepository.findAll();
     }
 
+    @Override
     public void addCourse(CourseModel courseModel) {
         courseRepository.save(courseModel);
     }
 
+    @Override
     public String editCourse(Integer courseId, CourseModel courseModelFromRequest) {
         Optional<CourseModel> coursemodel = courseRepository.findBycourseId(courseId);
         if (coursemodel.isPresent()) {
@@ -100,6 +115,7 @@ public class AdminService {
         return "Course not found";
     }
 
+    @Override
     public String deleteCourse(Integer id) {
         Optional<CourseModel> coursemodel = courseRepository.findById(id);
         if (coursemodel.isPresent()) {
@@ -108,7 +124,8 @@ public class AdminService {
         }
         return "Course not Found";
     }
-
+    
+    @Override
     public CourseModel getCourse(Integer id) {
         Optional<CourseModel> coursemodel = courseRepository.findById(id);
         if (coursemodel.isPresent()) {
@@ -117,14 +134,17 @@ public class AdminService {
         return null;
     }
 
+    @Override
     public List<CourseModel> getCourse() {
         return courseRepository.findAll();
     }
 
+    @Override
     public void addInstitute(InstituteModel instituteModel) {
         instituteRepository.save(instituteModel);
     }
 
+    @Override
     public String editInstitute(Integer instituteId, InstituteModel updatedInstitute) {
         Optional<InstituteModel> institutemodel = instituteRepository.findByInstituteId(instituteId);
         if (institutemodel.isPresent()) {
@@ -142,6 +162,7 @@ public class AdminService {
         return "Institute not found";
     }
 
+    @Override
     public String deleteInstitute(Integer id) {
         Optional<InstituteModel> institutemodel = instituteRepository.findByInstituteId(id);
         if (institutemodel.isPresent()) {
@@ -151,6 +172,7 @@ public class AdminService {
         return "Institute not Found";
     }
 
+    @Override
     public InstituteModel getInstitute(Integer id) {
         Optional<InstituteModel> institutemodel = instituteRepository.findByInstituteId(id);
         if (institutemodel.isPresent()) {
@@ -159,10 +181,12 @@ public class AdminService {
         return null;
     }
 
+    @Override
     public List<InstituteModel> getInstitutes() {
         return instituteRepository.findAll();
     }
 
+    @Override
     public String addReason(AdmissionModel admissionModel) {
         Optional<AdmissionModel> admissionModel1 = admissionRepo
                 .findByStudentIdNumberAndCourseId(admissionModel.getStudentIdNumber(), admissionModel.getCourseId());
@@ -174,10 +198,12 @@ public class AdminService {
         return "Course not Found";
     }
 
+    @Override
     public ResponseEntity<?> getAdmission(int page, int pageSize) {
         return ResponseEntity.ok(admissionRepo.findAll(PageRequest.of(page, pageSize)));
     }
 
+    @Override
     public ResponseEntity<?> filterByStatus(String status) {
         List<AdmissionModel> admission = admissionRepo.findByStatus(status);
         if (admission.isEmpty()) {
